(svn r14269) [0.6] -Backport from trunk: - Fix: Merge keycode for "normal" 0-9 keys and keypad 0-9 keys so people do not get confused that the keypad does not work as expected [FS#2277] (r14260) - Fix: Clicking on the smallmap didn't break the "follow vehicle in main viewport" [FS#2269] (r14243) - Fix: The engine-purchase-list-sorter doubled running-cost and halfed capacity of double-headed engines [FS#2267] (r14239) - Fix: Feeder share was computed wrong when splitting cargo packet (r14234) - Fix: Signs (town name, station name, ...) could be too long for 8bit width in pixels (r14221) - Fix: 10 days != 6*2.5 days, effectively causing the payment graph to show the wrong data (r14219) - Fix: When determining length of a string with limited size, first check if we are not out of bounds already (r14204)

+++ src/openttd.h
@@ -115,7 +115,7 @@ enum {
 struct ViewportSign {
 	int32 left;
 	int32 top;
-	byte width_1, width_2;
+	uint16 width_1, width_2;
 };
 
 enum {
